% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{basicDM}
\alias{basicDM}
\title{basicDM}
\usage{
basicDM(roads, car, packages)
}
\arguments{
\item{roads}{See help documentation for the runDeliveryMan function}

\item{packages}{See help documentation for the runDeliveryMan function}

\item{cars}{See help documentation for the runDeliveryMan function}
}
\value{
See help documentation for the runDeliveryMan function
}
\description{
This control function will pick up and deliver the packages in the order they
are given (FIFO). The packages are then delivered ignoring the trafic conditions
by first moving horizontally and then vertically.
}
\details{
As a first step, you should make sure you do better than this.
}
